feat(booking): Phase 3.2 — negotiation loop (submit, accept, modify, reject, customer decide)
Implements the full vendor negotiation lifecycle on top of the Phase 3.1
draft-items foundation. Vendors can accept, propose modifications with
diff_snapshot, or reject. Customers review diffs and accept/reject. All
transitions are idempotent, snapshotted, and fully logged.

- Enums: ModificationStatus, ModificationProposalKind, ModificationChangeKind
- Migrations: booking_modifications, booking_modification_items, idempotency_keys
- Models: BookingModification, BookingModificationItem (append-only items)
- Events: BookingSubmittedToVendor, VendorAccepted, VendorModificationProposed,
  VendorRejected, CustomerModificationDecided, BookingConfirmed, BookingCancelled
- Actions: SubmitBookingAction, VendorAcceptBookingAction, VendorModifyBookingAction,
  VendorRejectBookingAction, CustomerConfirmModifiedBookingAction (with idempotency)
- Listeners: WriteNegotiationSnapshotListener, ConfirmInventoryReservationsListener,
  ReleaseInventoryOnCancellationListener
- HTTP: BookingNegotiationController (customer), Vendor/BookingController,
  all FormRequests, BookingModificationResource
- Filament: BookingsMonitorResource (read-only, vendor_review + customer_review filter)
- Tests: 77/77 Pest tests passing across 7 test files (162 assertions)
- PHPStan: Booking module clean (added @return generics, @property docblocks)

+++ app/Modules/Booking/Http/Controllers/Customer/BookingItemController.php
@@ -20,7 +20,7 @@ class BookingItemController
         if ($booking === null) {
             return ApiResponse::error('Booking not found', 404);
         }
-        $item = app(AddItemToBookingAction::class)->execute($request->toDTO($booking->id, auth()->id()));
+        $item = app(AddItemToBookingAction::class)->execute($request->toDTO($booking->id, (int) auth()->id()));
 
         return ApiResponse::success(new BookingItemResource($item), [], 201);
     }
@@ -31,7 +31,7 @@ class BookingItemController
         if ($booking === null) {
             return ApiResponse::error('Booking not found', 404);
         }
-        $result = app(RemoveItemFromBookingAction::class)->execute($booking->id, auth()->id(), $itemPublicId);
+        $result = app(RemoveItemFromBookingAction::class)->execute($booking->id, (int) auth()->id(), $itemPublicId);
 
         return ApiResponse::success($result);
     }
